Build a map of geoTIFF of predictions for specified model fit

This function may be run at the console, but it's typically spun off as a batch job on Unity by map.

Usage

do_map(
  fit,
  clip = NULL,
  sourcedir = the$flightsdir,
  result = NULL,
  cores = 1,
  runinfo = NULL
)

Arguments

fit

Model fit object

clip

Optional clip, vector of xmin, xmax, ymin, ymax

sourcedir

Source directory, probably the flights directory

result

Result path and filename, sans extension

cores

Number of CPU cores to use

runinfo

Path and filename of run info file. When the run finishes, run info is written to this file.

Details

Side effects:

writes a geoTIFF, <result>.tif with, and a run info file

<runinfo>.RDS, with the following:

Time taken for the run (s)

Maximum memory used (GB)

Raster size (M pixel)

R error, or NULL for success

Requires rasterPrep. Install it with: remotes::install_github('ethanplunkett/rasterPrep')