Produce geoTIFF maps of predicted vegetation cover from fitted models

Console command to launch a prediction run via do_map, typically in a batch job on Unity.

Usage

do_map(site, fitid, fitfile, clip, result, mapid, rep = NULL)

Arguments

site

Three letter site code. If fitting from a fit id that was built on a single site, you may omit site to map the same site (this is the most common situation). If you want to map sites other than the site the model was built on, or the model was built on mutiple sites, site is required.

fitid

id in the fits database (NULL if not specified)

fitfile

Full specification of an RDS file with the fit object

clip

Optional clip, vector of xmin, xmax, ymin, ymax

result

Result file name

mapid

Id in maps database

rep

Throwaway argument to make slurmcollie happy

Details

Side effects:

writes a geoTIFF, <result>.tif with, and a run info file

<runinfo>.RDS, with the following:

Time taken for the run (s)

Maximum memory used (GB)

Raster size (M pixel)

R error, or NULL for success

Requires rasterPrep. Install it with: remotes::install_github('ethanplunkett/rasterPrep')